Add in-app tutorial, fix compilations issues w/ legacy libs.
In order to use the viewpager2 dependency, I needed to enable AndroidX,
which in turn made it necessary to stop using some legacy UI dependencies.
Now RECORDA has more or less the same dependencies (apart from libxdf and
the Movella SDK) as SENDA, which makes future development and migration to
Kotlin far easier.